[WIP][TASK] Extract VariantsProcessor from SearchResultSetService
This pr:

* Extracts the creation of the variants into a SearchResultSetProcessor
* Reduces the complexity of SearchResultSetService
